requirements.txt
PyQt5==5.15.9
flask==3.0.3
termcolor==3.3.0
PyJWT==2.9.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
flask_sqlalchemy==3.1.1
pillow==12.2.0
uuid==1.30
dotenv==0.9.9
langchain==1.2.15
langchain-core==1.2.28
langchain-openai==1.1.12
tiktoken==0.12.0
Werkzeug==3.1.8
